requirements.txt
# Core application dependencies
fastapi==0.124.4
litellm==1.80.11
openai-agents
python-dotenv==1.2.1
requests==2.32.5
PyJWT==2.10.1

# FastAPI and server dependencies
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
uvicorn==0.38.0
anyio==4.12.0
typing_extensions==4.15.0
annotated-types==0.7.0

# HTTP client dependencies
urllib3==2.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
